requirements.txt
Flask
Flask-SQLAlchemy
python-docx
PyMuPDF
